Updated logging dependency versions and lowered level of logging for Spring
Updated versions of logging dependencies to latest version. The other
logback version appeared to be missing a class.
Also lowered the level of logging for Spring, as the debug statements
were clogging the console while I was trying to debug things. Will
reconfigure it to go somewhere more useful than the console eventually.