Experimenting with logger names and file output
Added an appender (that is, a definition of where the logs go) to send
logs to a file. It's configured at the moment with a rolling file
appender to archive files each day and start a new log file.
You can give each logger any name, or specify a package instead.