Updated from 3-30 meeting
It's registering everything as new or some reason? I guess I changed
around quite a few things.
-Split the main exception into multiple exceptions and
adjusted unit testing accordingly
-Cleaned up the backtracking detection
-Moved turn switching into the move method and adjusted the unit tests
accordingly. I had to add in a constructor that takes in a 2D array
giving the board's starting state to avoid having to make a bunch of
extraneous moves for the player I wasn't testing for.
- Added a new unit test for a win condition (forgot that one).
Currently, you have to query the board to see if it's in a winning
state.
- Added in player IDs for player one and two and adjusted everything to
work on those. Note that the *pieces* still need the Player enum, and
it's sometimes more convenient to use that the IDs in the private
methods, so it's still in various places on the board as well. I
considered putting the IDs into an array, but then technically you
have...player 0 and player 1, which is weird, but could work. It's fine
as is for now though, I think. It'd just be easier to extend to a third
and fourth player if it were an array.