SQL and gradle build update for database with login
Added dependencies needed (well, not positive they're all needed, but
they were in the tutorial - we can take them out if we dont use them)
for using a database with Spring Security login.
Also updated the MySQL with an extra row in the user table (Enabled? Not
sure how it's used yet) and a user_roles table that will be needed to
snag the role for the Spring login. NOTE: I did update the table named
"user" to "users". I can change it back if necessary.