Edits in response to FindBugs
FindBugs suggests not exposing mutable objects. It seems a good idea to
me, but I've separated out each FindBugs recommendation in case you
think it's not worth it and we should revert the change.
Let me know if there's a better way to copy the 2D array...given that we
know the size this seemed the clearest.