Logging updates 2
Updated logging in the GameRestController. Mostly it is paralleling what
goes into the audit logs, because it seemed like it should. I also added
an audit log of the starting lines when a game is started. The reason
for this is that in the audit logs we have the piece ID and the
destinations, but not the starting location. This makes it hard to know
where a piece came from when the starting line is shuffled.